Implement audio: OPL music, AdLib/PC speaker SFX, digitized sounds
Replace stubbed sound manager with full I2S audio implementation:

- DMA IRQ fill callback generates OPL music (IMF sequencer at 700 Hz),
  AdLib sound effects (140 Hz), PC speaker square waves, and mixes up
  to 6 digitized sound channels with stereo panning
- SD_PrepareSound resamples 7042 Hz digi data to 44100 Hz in PSRAM
- SD_PlayDigitized assigns sounds to mixer channels with L/R volume
- SD_SoundPlaying/SD_WaitSoundDone now check actual playback state

Fix MAME fmopl to output raw 32-bit mono samples — the previous code
wrote stereo pairs (buf[i*2]) but callers read mono (buf[i]), halving
the effective sample rate and causing aliasing heard as metallic sound.
Removed the internal 4x boost and 16-bit clamp; scaling is now handled
cleanly in the mixer with full 32-bit dynamic range.

+++ src/wolf3d/mame/fmopl.h
@@ -48,7 +48,7 @@ int  YM3812Write(int which, int a, int v);
 unsigned char YM3812Read(int which, int a);
 void YM3812Mute(int which,int channel,BOOL mute);
 int  YM3812TimerOver(int which, int c);
-void YM3812UpdateOne(int which, INT16 *buffer, int length);
+void YM3812UpdateOne(int which, INT32 *buffer, int length);
 
 void YM3812SetTimerHandler(int which, OPL_TIMERHANDLER TimerHandler, int channelOffset);
 void YM3812SetIRQHandler(int which, OPL_IRQHANDLER IRQHandler, int param);
